# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prayag Polymers pricing-intelligence API (product catalog, MRP history, competition analysis)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /catalog/products:
    get:
      operationId: getCatalogProducts
      tags: [catalog]
      summary: List catalog products with current MRP
      parameters:
        - name: division
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: priceStatus
          in: query
          required: false
          description: Filter by price presence. "priced" or "pending".
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
        - name: asOf
          in: query
          required: false
          description: Resolve catalog availability and MRP as of this date (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogProductList"
  /catalog/products/{itemCode}:
    get:
      operationId: getCatalogProduct
      tags: [catalog]
      summary: Get a product with its full MRP price history
      parameters:
        - name: itemCode
          in: path
          required: true
          schema:
            type: string
        - name: asOf
          in: query
          required: false
          description: Resolve product availability as of this date (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Product detail with history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /catalog/products/{itemCode}/mrp:
    patch:
      operationId: patchCatalogProductMrp
      tags: [catalog]
      summary: Set a new current MRP for a product (inline edit, no file upload required)
      parameters:
        - name: itemCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mrp]
              properties:
                mrp:
                  type: number
                  description: New MRP value (must be positive)
                effectiveDate:
                  type: string
                  description: "Effective date (YYYY-MM-DD). Defaults to today."
                notes:
                  type: string
                  description: Optional note for the audit trail
      responses:
        "200":
          description: MRP updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, itemCode, mrp, effectiveDate]
                properties:
                  ok:
                    type: boolean
                  itemCode:
                    type: string
                  mrp:
                    type: number
                  effectiveDate:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    CatalogProductListItem:
      type: object
      required: [id, itemCode, hasPrice]
      properties:
        id:
          type: integer
        itemCode:
          type: string
        productName:
          type: ["string", "null"]
        division:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        seriesRange:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        uom:
          type: ["string", "null"]
        isActive:
          type: boolean
        dataFlag:
          type: ["string", "null"]
        discontinuedFrom:
          type: ["string", "null"]
          format: date
        hasPrice:
          type: boolean
        currentMrp:
          type: ["number", "null"]
        currentNet:
          type: ["number", "null"]
        currentBasis:
          type: ["string", "null"]
        effectiveDate:
          type: ["string", "null"]
    CatalogProductList:
      type: object
      required: [rows, total, page, pageSize]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/CatalogProductListItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    CatalogProduct:
      type: object
      required: [id, itemCode]
      properties:
        id:
          type: integer
        itemCode:
          type: string
        productName:
          type: ["string", "null"]
        division:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        seriesRange:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        uom:
          type: ["string", "null"]
        kgCost:
          type: ["number", "null"]
        isActive:
          type: boolean
        sourceFiles:
          type: ["string", "null"]
        dataFlag:
          type: ["string", "null"]
        discontinuedFrom:
          type: ["string", "null"]
          format: date
    MrpPriceRow:
      type: object
      required: [id, itemCode, effectiveDate, loadDate, priceBasis, isCurrent]
      properties:
        id:
          type: integer
        itemCode:
          type: string
        mrp:
          type: ["number", "null"]
        netPrice:
          type: ["number", "null"]
        discountPct:
          type: ["number", "null"]
        priceBasis:
          type: string
        effectiveDate:
          type: string
        loadDate:
          type: string
        sourceFile:
          type: ["string", "null"]
        isCurrent:
          type: boolean
        notes:
          type: ["string", "null"]
        reviewStatus:
          type: string
        reviewReasons:
          type: ["string", "null"]
        importBatchId:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
          format: date-time
    CatalogProductDetail:
      type: object
      required: [product, history]
      properties:
        product:
          $ref: "#/components/schemas/CatalogProduct"
        history:
          type: array
          items:
            $ref: "#/components/schemas/MrpPriceRow"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key